requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deep-translator==1.11.4
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==4.7.2
python-dotenv==1.1.1
PyYAML==6.0.1
requests==2.32.3
soupsieve==2.6
ua-parser==0.18.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
Werkzeug==3.1.3
spacy==3.7.4
